Fix conditional execution of `ci-ok` for community PRs (#16573)
In #16567 we added a step for skipping `ci-ok` for community PRs (since
these PRs have their acceptance tests run in a workflow separate to the
PR being checked). However, we didn't modify the default step not to run
in this case, so we still see failures for community PRs. This commit
fixes that.